backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.7
qdrant-client==1.7.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0